Move comparator and hasher classes to string compare
