+++ src/apps/ENGINE/s_functab.h
@@ -2,7 +2,7 @@
 
 #include "s_import_func.h"
 #include "s_vartab.h"
-#include "strutils.h"
+#include "storm/string_compare.hpp"
 #include <unordered_map>
 #include <vector>
 
@@ -94,7 +94,6 @@ class FuncTable
 
   private:
     std::vector<FuncInfo> funcs_;
-    CaseInsensitiveStringHasher hasher_;
-    CaseInsensitiveStringComparator comparator_;
-    std::unordered_map<std::string, size_t, CaseInsensitiveStringHasher, CaseInsensitiveStringComparator> hash_table_;
+    storm::iStrHasher hasher_;
+    std::unordered_map<std::string, size_t, storm::iStrHasher, storm::iStrComparator> hash_table_;
 };
